# Use SpecialEnv + smaller testing db to speed up some slow BackupEngineRateLimitingTestWithParam