--- a/utilities/backup/backup_engine_impl.h
+++ b/utilities/backup/backup_engine_impl.h
@@ -25,5 +25,12 @@ struct TEST_BackupMetaSchemaOptions {
 void TEST_SetBackupMetaSchemaOptions(
     BackupEngine *engine, const TEST_BackupMetaSchemaOptions &options);
 
+// Modifies the BackupEngine(Impl) to use specified clocks for backup and
+// restore rate limiters created by default if not specified by users for
+// test speedup.
+void TEST_SetDefaultRateLimitersClock(
+    BackupEngine* engine,
+    const std::shared_ptr<SystemClock>& backup_rate_limiter_clock = nullptr,
+    const std::shared_ptr<SystemClock>& restore_rate_limiter_clock = nullptr);
 }  // namespace ROCKSDB_NAMESPACE
 #endif  // ROCKSDB_LITE