@hx235 2022-05-10
> My main question/concern is whether part of the code will use a special env over Default and other code will use env over something else. Going through doing something similar on other changes in the past, I found this can lead to weird errors.

Good point - I was trying to replace as minimum env as possible but didn't foresee how this inconsistency might lead us to problem (and probably readability issue as reader might be in doubt). Let me think about it more. Thanks!