@hx235 2022-05-12
> > where its clock accepts fake elapse of time during rate-limiting (6582 - 28560 ms/per test)
> 
> Is it still waiting for something? Or do these tests do tens of seconds of actual work?

Most of the time went into backing-up/restoring the db - shrunk the db by 10x.